Tightens the public API provided by the events module.
