+++ tools/cli/CMakeLists.txt
@@ -10,10 +10,10 @@ endif()
 
 # The public executable.
 add_executable(ion main.cpp cli.cpp)
-target_include_directories(ion PRIVATE ./ ../events docopt)
+target_include_directories(ion PRIVATE ./ ../events ../events/inc docopt)
 target_link_libraries(ion ionc ion_events docopt)
 
 # A library for testing.
 add_library(ion_cli cli.cpp)
-target_include_directories(ion_cli PRIVATE ./ ../events)
+target_include_directories(ion_cli PRIVATE ./ ../events ../events/inc)
 target_link_libraries(ion_cli ionc ion_events)
