+++ tools/events/ion_event_equivalence.cpp
@@ -14,35 +14,16 @@
 
 #include <set>
 #include "ion_event_equivalence.h"
+#include "ion_event_equivalence_impl.h"
 #include "ion_event_util.h"
+#include "ion_event_stream_impl.h"
 #include <ion_const.h>
 #include <cstdlib>
 #include <sstream>
 #include <ion_internal.h>
+#include "floating_point_util.h"
 
 TIMESTAMP_COMPARISON_FN g_TimestampEquals = ion_timestamp_equals;
-
-#define ION_EVENT_EQUIVALENCE_PARAMS \
-IonEventStream *ION_STREAM_EXPECTED_ARG, size_t ION_INDEX_EXPECTED_ARG, IonEventStream *ION_STREAM_ACTUAL_ARG, \
-size_t ION_INDEX_ACTUAL_ARG, ION_EVENT_COMPARISON_TYPE ION_COMPARISON_TYPE_ARG, IonEventResult *ION_RESULT_ARG
-
-#define ION_EVENT_EQUIVALENCE_ARGS \
-ION_STREAM_EXPECTED_ARG, ION_INDEX_EXPECTED_ARG, \
-ION_STREAM_ACTUAL_ARG, ION_INDEX_ACTUAL_ARG, ION_COMPARISON_TYPE_ARG, ION_RESULT_ARG
-
-#define ION_GET_EXPECTED ION_STREAM_EXPECTED_ARG->at(ION_INDEX_EXPECTED_ARG)
-#define ION_GET_ACTUAL ION_STREAM_ACTUAL_ARG->at(ION_INDEX_ACTUAL_ARG)
-#define ION_SET_EXPECTED IonEvent *ION_EXPECTED_ARG = ION_GET_EXPECTED
-#define ION_SET_ACTUAL IonEvent *ION_ACTUAL_ARG = ION_GET_ACTUAL
-#define ION_PREPARE_COMPARISON ION_SET_EXPECTED; ION_SET_ACTUAL
-#define ION_NEXT_EXPECTED_INDEX ION_INDEX_EXPECTED_ARG++
-#define ION_NEXT_ACTUAL_INDEX ION_INDEX_ACTUAL_ARG++
-#define ION_NEXT_INDICES ION_NEXT_EXPECTED_INDEX; ION_NEXT_ACTUAL_INDEX
-#define ION_ACTUAL_VALUE_LENGTH ion_event_value_length(ION_STREAM_ACTUAL_ARG, ION_INDEX_ACTUAL_ARG)
-#define ION_EXPECTED_VALUE_LENGTH ion_event_value_length(ION_STREAM_EXPECTED_ARG, ION_INDEX_EXPECTED_ARG)
-#define ION_NEXT_ACTUAL_VALUE_INDEX ION_INDEX_ACTUAL_ARG += ION_ACTUAL_VALUE_LENGTH
-#define ION_NEXT_EXPECTED_VALUE_INDEX ION_INDEX_EXPECTED_ARG += ION_EXPECTED_VALUE_LENGTH
-#define ION_NEXT_VALUE_INDICES ION_NEXT_EXPECTED_VALUE_INDEX; ION_NEXT_ACTUAL_VALUE_INDEX
 
 /**
  * Tests the values starting at the given indices in the given streams (they may be the same) for equivalence
